audio.c: unsigned int functions return unsigned ints, not size_t
git-svn-id: https://svn.musicpd.org/mpd/trunk@7208 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/audio.c
@@ -55,7 +55,7 @@ static mpd_sint32 audioBufferPos;
 
 unsigned int audio_device_count(void)
 {
-	size_t nr = 0;
+	unsigned int nr = 0;
 	ConfigParam *param = NULL;
 
 	while ((param = getNextConfigParam(CONF_AUDIO_OUTPUT, param)))
